Replace deprecated AsyncTask with ExecutorService
First try at replacing SendMessage which derives from AsyncTask (which is deprecated as of API level 30). Seems to work fine!
Leaving in the commented out stuff and new debug code for this initial commit. After testing with both tablets, both interfaces, and both locations, will clean it up and submit.